Shared_Cross_Modal_IE/requirements.txt
torch>=1.13.0
transformers>=4.30.0
Pillow>=9.0.0
torchcrf>=1.1.0
scikit-learn>=1.0.0
tqdm>=4.60.0
numpy>=1.20.0
matplotlib>=3.5.0
jupyter>=1.0.0
ipykernel>=6.0.0
